renames module to area_accounting
