+++ mescal/energy_data_handling/area_accounting/area_model_generator.py
