+++ mescal/energy_data_handling/area_accounting/area_variable_base.py
