+++ mescal/energy_data_handling/area_accounting/border_model_geometry_calculator.py
