+++ mescal/energy_data_handling/area_accounting/border_variable_capacity.py
@@ -4,7 +4,7 @@ from dataclasses import dataclass
 import numpy as np
 import pandas as pd
 
-from mescal.energy_data_handling.line_flow_data import NetworkLineCapacitiesData
+from mescal.energy_data_handling.network_lines_data import NetworkLineCapacitiesData
 from mescal.energy_data_handling.area_border.border_variable_base import AreaBorderVariableCalculatorBase
